Add files via upload
This is a responsive website which I created during my MSc studies. The aim was to create a database driven, website where the users can upload media related to business concept ("Make your own" clothing tutorials) However, the main focus is for website visitors to be able to request help and other to respond with video tutorials. The uploaded media must also include metadata and the completed website must allow for internal review by the website administrator to confirm the content is appropriate before releasing it for public consumption.Once the media is live the website must allow for user comments which also need to be approved by the administrator before their release. The website content should be displayed dynamically and organized based on the submitted metadata and filtered based on website users' requested subject master.